+++ include/bitcal/bit_block.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "config.hpp"
+
 #include "bit_view.hpp"
 
 #include <cassert>